# modules/sample-rest-builder/sample-rest-builder-impl/src/main/java/sample/rest/builder/internal/resource/v1_0/BaseBarResourceImpl.java
package sample.rest.builder.internal.resource.v1_0;

import com.liferay.petra.function.UnsafeFunction;
import com.liferay.portal.kernel.service.GroupLocalService;
import com.liferay.portal.kernel.service.ResourceActionLocalService;
import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
import com.liferay.portal.kernel.service.RoleLocalService;
import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
import com.liferay.portal.vulcan.util.TransformUtil;

import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.tags.Tags;

import java.util.List;

import javax.annotation.Generated;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import javax.validation.constraints.NotNull;

import javax.ws.rs.Consumes;
import javax.ws.rs.DELETE;
import javax.ws.rs.GET;
import javax.ws.rs.PATCH;
import javax.ws.rs.POST;
import javax.ws.rs.PUT;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.core.UriInfo;

import sample.rest.builder.dto.v1_0.Bar;
import sample.rest.builder.resource.v1_0.BarResource;

/**
 * @author me
 * @generated
 */
@Generated("")
@Path("/v1.0")
public abstract class BaseBarResourceImpl implements BarResource {

	/**
	 * Invoke this method with the command line:
	 *
	 * curl -X 'POST' 'http://localhost:8080/o/sample-rest-builder/v1.0/bars' -d $'{"active": ___, "description": ___, "name": ___}' --header 'Content-Type: application/json' -u 'test@liferay.com:test'
	 */
	@Override
	@Consumes({"application/json", "application/xml"})
	@POST
	@Path("/bars")
	@Produces({"application/json", "application/xml"})
	@Tags(value = {@Tag(name = "Bar")})
	public Bar postBar(Bar bar) throws Exception {
		return new Bar();
	}

	/**
	 * Invoke this method with the command line:
	 *
	 * curl -X 'DELETE' 'http://localhost:8080/o/sample-rest-builder/v1.0/bars/{barId}'  -u 'test@liferay.com:test'
	 */
	@Override
	@DELETE
	@Parameters(value = {@Parameter(in = ParameterIn.PATH, name = "barId")})
	@Path("/bars/{barId}")
	@Produces({"application/json", "application/xml"})
	@Tags(value = {@Tag(name = "Bar")})
	public Bar deleteBar(
			@NotNull @Parameter(hidden = true) @PathParam("barId") Long barId)
		throws Exception {

		return new Bar();
	}

	/**
	 * Invoke this method with the command line:
	 *
	 * curl -X 'GET' 'http://localhost:8080/o/sample-rest-builder/v1.0/bars/{barId}'  -u 'test@liferay.com:test'
	 */
	@Override
	@GET
	@Parameters(value = {@Parameter(in = ParameterIn.PATH, name = "barId")})
	@Path("/bars/{barId}")
	@Produces({"application/json", "application/xml"})
	@Tags(value = {@Tag(name = "Bar")})
	public Bar getBar(
			@NotNull @Parameter(hidden = true) @PathParam("barId") Long barId)
		throws Exception {

		return new Bar();
	}

	/**
	 * Invoke this method with the command line:
	 *
	 * curl -X 'PATCH' 'http://localhost:8080/o/sample-rest-builder/v1.0/bars/{barId}' -d $'{"active": ___, "description": ___, "name": ___}' --header 'Content-Type: application/json' -u 'test@liferay.com:test'
	 */
	@Override
	@Consumes({"application/json", "application/xml"})
	@PATCH
	@Parameters(value = {@Parameter(in = ParameterIn.PATH, name = "barId")})
	@Path("/bars/{barId}")
	@Produces({"application/json", "application/xml"})
	@Tags(value = {@Tag(name = "Bar")})
	public Bar patchBar(
			@NotNull @Parameter(hidden = true) @PathParam("barId") Long barId,
			Bar bar)
		throws Exception {

		Bar existingBar = getBar(barId);

		if (bar.getActive() != null) {
			existingBar.setActive(bar.getActive());
		}

		if (bar.getDateCreated() != null) {
			existingBar.setDateCreated(bar.getDateCreated());
		}

		if (bar.getDateModified() != null) {
			existingBar.setDateModified(bar.getDateModified());
		}

		if (bar.getDescription() != null) {
			existingBar.setDescription(bar.getDescription());
		}

		if (bar.getName() != null) {
			existingBar.setName(bar.getName());
		}

		return putBar(barId, existingBar);
	}

	/**
	 * Invoke this method with the command line:
	 *
	 * curl -X 'PUT' 'http://localhost:8080/o/sample-rest-builder/v1.0/bars/{barId}' -d $'{"active": ___, "description": ___, "name": ___}' --header 'Content-Type: application/json' -u 'test@liferay.com:test'
	 */
	@Override
	@Consumes({"application/json", "application/xml"})
	@PUT
	@Parameters(value = {@Parameter(in = ParameterIn.PATH, name = "barId")})
	@Path("/bars/{barId}")
	@Produces({"application/json", "application/xml"})
	@Tags(value = {@Tag(name = "Bar")})
	public Bar putBar(
			@NotNull @Parameter(hidden = true) @PathParam("barId") Long barId,
			Bar bar)
		throws Exception {

		return new Bar();
	}

	public void setContextAcceptLanguage(AcceptLanguage contextAcceptLanguage) {
		this.contextAcceptLanguage = contextAcceptLanguage;
	}

	public void setContextCompany(
		com.liferay.portal.kernel.model.Company contextCompany) {

		this.contextCompany = contextCompany;
	}

	public void setContextHttpServletRequest(
		HttpServletRequest contextHttpServletRequest) {

		this.contextHttpServletRequest = contextHttpServletRequest;
	}

	public void setContextHttpServletResponse(
		HttpServletResponse contextHttpServletResponse) {

		this.contextHttpServletResponse = contextHttpServletResponse;
	}

	public void setContextUriInfo(UriInfo contextUriInfo) {
		this.contextUriInfo = contextUriInfo;
	}

	public void setContextUser(
		com.liferay.portal.kernel.model.User contextUser) {

		this.contextUser = contextUser;
	}

	protected <T, R> List<R> transform(
		java.util.Collection<T> collection,
		UnsafeFunction<T, R, Exception> unsafeFunction) {

		return TransformUtil.transform(collection, unsafeFunction);
	}

	protected <T, R> R[] transform(
		T[] array, UnsafeFunction<T, R, Exception> unsafeFunction,
		Class<?> clazz) {

		return TransformUtil.transform(array, unsafeFunction, clazz);
	}

	protected <T, R> R[] transformToArray(
		java.util.Collection<T> collection,
		UnsafeFunction<T, R, Exception> unsafeFunction, Class<?> clazz) {

		return TransformUtil.transformToArray(
			collection, unsafeFunction, clazz);
	}

	protected <T, R> List<R> transformToList(
		T[] array, UnsafeFunction<T, R, Exception> unsafeFunction) {

		return TransformUtil.transformToList(array, unsafeFunction);
	}

	protected AcceptLanguage contextAcceptLanguage;
	protected com.liferay.portal.kernel.model.Company contextCompany;
	protected HttpServletRequest contextHttpServletRequest;
	protected HttpServletResponse contextHttpServletResponse;
	protected Object contextScopeChecker;
	protected UriInfo contextUriInfo;
	protected com.liferay.portal.kernel.model.User contextUser;
	protected GroupLocalService groupLocalService;
	protected ResourceActionLocalService resourceActionLocalService;
	protected ResourcePermissionLocalService resourcePermissionLocalService;
	protected RoleLocalService roleLocalService;

}